requirements.txt
pandas
pydantic
fastapi
uvicorn
python-multipart
aiofiles
PyYAML
paho-mqtt
python-dotenv